Cretae
Calculates the system's energy production efficiency as a percentage for every hour. The formula is based on a ratio between the absolute values of power output and input, respectively. This computation is only relevant for the charging and discharging state. All of the outputted values are stored in the eff.csv file.